requirements.txt
Django~=4.1.7
factory-boy==3.2.1
gunicorn~=20.1.0
